rectif real

+++ src/engine/expression/litterales/lreal.h
@@ -7,13 +7,12 @@
 namespace Engine {
 class Lreal : public Lnumerical, public R1value
 {
-    float real;
+    double real;
 public:
     Lreal(const std::string s);
-    Lreal(float r): Lnumerical("Lreal"),real(r){};
-    Lreal* setReal(float r);
+    Lreal(double r): Lnumerical("Lreal"),real(r){};
+    Lreal* setReal(double r);
     double getValue() override {return real;}
-//    LReal* createCopy()const override{return new LReal(*this);}
     std::string const toString()const override{return std::to_string(real);}
     Expression* getCopy() const override { return new Lreal(real); }
     Expression* simplifyType() override;
